MSM/Serialization: move main function in interpreter
Fix https://github.com/o1-labs/proof-systems/issues/1890
